parser: use precedence climbing algorithm
This changes the parser to use precedence climbing [1] instead of
Jon Blow's rewrite rules.

[1]: http://eli.thegreenplace.net/2012/08/02/parsing-expressions-by-precedence-climbing

Precedence climbing is better since it avoids a linear loop inside the
binary operator parsing (to find the right place to insert the new node)
and it also naturally fixes the precedence of unary operators (which
previously would potentially consume all the rest of the document if it
happened to be a valid expression).