parser: disallow trailing operators for :=, +, etc.
Without this, strings like "x : = 1" would parse the same as "x = 1"
because the ":" was considered a trailing operator (and discarded).

We use trailing operators to allow constructs like "(1, 2,)" and
"foo(); bar();" (where the trailing operators are "," and ";",
respectively.)

This also gets rid of a special case for juxtaposition, since we can
now just parse it without allow trailing operators (which in the case
of juxtaposition is the zero string).