value: new class representing run-time variables
In order to start compiling, we need something to represent run-time
variables at compile time.

Here a value's metatype is either global, local, or constant. We may add
more metatypes in the future, but for now, they mean:

 - global: a value with a specific constant address (to dereference it,
   you just have to dereference the address)

 - local: a value local to a given stack frame (to dereference it, you
   have to add it as an offset to the current stack pointer)

 - constant: a constant value

We also have to start sketching out a type system. For now, we
distinguish between types of values (void, boolean, int, etc.), but we
don't really know anything about the types themselves (sizes, members,
etc.).

We do need to have some built-in types, but we should try to keep the
base language small.