value: work towards better types
I'm not 100% sure about the direction this is taking. We need a "type"
type and we need a way to convert integer literals to uint64 values.
This requires some sort of constructor, but I'm not sure if it should
be a function or a macro (but almost certainly the latter) and it's a
bit ugly to make 'struct value_type' start depending on the function,
scope, and value structs (since that introduces a recursive dependency
of sorts).

Anyway, this is what I have for now.