main: factor out run() function
I want "@foo" (AKA "(_eval (symbol foo))") to evaluate "foo" at compile
time.

Instead of duplicating all the logic for compiling and evaluating an
expression, I'd like to implement evaluation as compilation + execution.

To this end, let's factor out the run() function so we can reuse it in
in the builtin "_eval" macro.