builtin/fun: make 'fun' macro return a type
The 'fun' macro used to take a function body (an AST node) directly, but
that makes it difficult to talk about function types (signatures).

So I've made 'fun' instead return a type -- the function type -- which is
itself a callable, taking the function body as an argument. Calling the
function type with a body compiles the body to machine code and returns a
pointer to the compiled function.

This now works:

    f := (fun () u64) {
        print u64 1;
        u64 1;
    };

    f();

This declares 'f' to be a function returning u64 and calls it.