builtin/if: change semantics
This changes semantics of the 'if' macro slightly so that its (return)
value is 'void' if there is no 'else' block or the two blocks have
different (return) types.

This is a trade-off between having an "imperative if" which doesn't
return anything and a "functional if" which returns the value of whatever
branch was taken.

Everything is, however, as usual, subject to change.