builtin/operators: generalise builtin operators
Types can specify their own operators, i.e. builtin_type_int and
builtin_type_u64 will have different functions as their ->add
members.

In general, operators like + or <= should evaluate the LHS (which
gives us both the value and the type of the LHS) and then call
type->operator with the (unevaluated) RHS.

Since all the operators work the same way, we can generalise the
existing code for "add" and use it for all the operators.