builtin/quote: new builtin macro
This is intended to work (barring bugs) like Lisp and Scheme's "quote"
operator. It returns a value with the type of "ast_node", thus allowing
you to pass syntax to functions at run-time (macros take their arguments
as syntax, but work at compile-time). You currently cannot do anything
with these values yet, but the intention is to provide "compile" and
"eval" functions which take arguments of this type.