function: reserve stack space for locals
Non-leaf function were not reserving space on the stack for its locals,
causing contained calls to potentially overwrite the locals and the return
address.

This switches function generation to use a sequence like this:

  // prologue
  pushq %rbp
  movq %rsp, %rbp
  pushq %rbx
  subq $x, %rsp
  ...
  // epilogue
  addq $x, %rsp
  popq %rbx
  popq %rbp
  retq

This can be done better (like most of our code generation), but works fine
for now.

Locals are also addressed as offsets from %rbp rather than %rsp.