compile: correctly copy out result of eval
We used to just return the result of compiling the evaled expression, but
this is wrong if the value is a local, since it would be local to the
function we called to do the actual evaluation and NOT local to the calling
function.

The solution is to just save the result in a value which is known to be
global before returning.