scope: disallow the use of locals outside the defining function
We should not allow locals from one function to be accessed from an inner
function (at least not directly without a lambda capture), since the
frame pointer offset of a local will only be valid for the function where
the local was defined.

The solution is to record which function defined each symbol and only
allow a symbol to be looked up if the function doing the looking up is
the same function as the defining function.

  @f := (fun u64 (u64)) (b) {
      a := u64 1;
      g := (fun u64 (u64)) (b) {
          a + b;
      };
      g(b);
  };

  print f(u64 2);

gives this error:

  snippets/fun-scopes.v:4:8: cannot access local of different function
          a + b;
          ^