scope: keep track of compile-time contexts
This one is a bit tricky.

The @ operator allows an expression to be evaluated at compile time.

When an expression is evaluated at compile time, it is first itself
compiled, then executed.

The @ operator can be nested, so the "run time" of one expression
can be the "compile time" of another (and vice versa).

An example is perhaps more illustrative:

  t := fun u64 ();
  t () u64 0;

Here, "t" is defined as a new type and then its constructor (which is
a macro) is invoked (at compile time). However, although we know the
_type_ of t when it is invoked, the definition/assignment of its
actual value happens when the whole block is _executed_, thus "t" does
not have a value at compile time. This causes the program/compiler to
crash as it dereferences "t" which is still uninitialised.

The patch adds tracking of compile time contexts and prevents us
from trying to evaluate variables which are not available at compile
time. Instead of crashing on the example above, the compiler now outputs
this message:

  tests/builtin/define-lookup-1.v:2:0: cannot access run-time variable t at compile time
  t () u64 0;
  ^