builtin/operators: move to members
Instead of having a fixed set of operators on a type, we make the operators
regular members with special names.

    x := u64 1;
    y := u64 2;

    x + y;

Here, 'x + y' is parsed as (juxtapose _add (juxtapose x y)), which will
call the builtin macro '_add' with 'x' and 'y' as arguments.

The builtin macro '_add' will evaluate the type of 'x', look up the
'_add' member of the type and call it.

At the moment, all members are function pointers which are called at
compile time, but in the future members could be any kind of macro (or
value).