compile: don't eval() the LHS of a juxtaposition
Okay, this one is going to take some effort to explain.

For compile_juxtapose(), we used to first compile the LHS to figure out
what type it was; if it was a type or a macro, we'd call eval() on the
same node (which compiles it a second time within a new temporary function).

Compiling the same expression twice is unfortunate because of potential
side effects (think C where you have an array of function pointers and you
do a[i++](); -- you wouldn't expect i++ to happen twice just because it's
somewhere in the LHS of a function call.

What I've done is to add context tracking to the compiled values themselves.
This context is where the value was defined. Remember that a value cannot be
used in any of the parents of the context where it was defined (see
tests/builtin/define-lookup-*.v).

Global values which are assigned immediately (e.g. the return value of
invoking 'fun') are created in the top-level context so it can be used
immediately by any context.