compile: factor out _compile_juxtapose() helper
This helper takes both an AST node and a value for the LHS and can be used
when the LHS has already been compiled.

This is useful for things like the _add macro, which rewrites e.g. (_add x y)
as (x._add y), without having to generate new AST nodes.