operators: fix (x._add y) syntax
Operators like + and < work like this:

In the parser, we parse (x + y) as (_add (x y)).

When this is compiled, we look up the name _add, which resolves to a builtin
macro.

This macro compiles x to get its type, then looks to see if the type has a
member called _add. Then it invokes this macro. This is equivalent to typing
(x._add y) directly.

Assuming the built-in macros are used, these three expressions should be
equivalent and evaluate to the same value.