builtin/fun: fix bug in calling void functions
When calling void functions, we should not allocate a local variable to
hold the return value after the function has returned. Doing this causes
us to attempt to allocate a stack slot with the type builtin_type_void,
which has 0 size and 0 alignment. In particular, aligning anything to 0
using (x + align - 1) & ~(align - 1) causes the result to be 0, which
means we will overwrite 0(%rbp) with %rax and corrupt the stack frame.