compile: wrap context/function/scope in a new type
We're passing around the tuple <context, function, scope> in a *lot* of
places.

These are shared_ptrs, which means we're doing a lot of copying (and the
corresponding bumping and un-bumping of reference counts). It is true
that we could pass them as references to pointers in most cases, which
would skip the reference counting, but the pointers would still take
up stack space.

Most of the time we don't even modify any of these from the caller to the
callee.

I wanted to add another field to the tuple, namely a reference to the
file being processed (so we don't have to add that to every single AST
node object), but passing all these arguments around is bordering on
madness.

What we can do instead is to factor these shared_ptrs out of the function
signature and into a new type.

The new type is passed by const reference (i.e. you're essentially only
passing around a pointer) so the overhead of passing it down to callees
is minimal. On the other hand, you are required to make an extra
dereference whenever you want to access any of its members. When you
want to _modify_ any of the members, you essentially create a new copy
on the stack and modify that. Then you pass the modified copy to the
callee.

This should be a significant save, if not in performance at least in
code readability.