ast: replace AST_PAIR nodes by calls to _declare
I was initially very doubtful about whether this was the right thing to do
or not -- having : as its own AST node type seems very natural.

HOWEVER, I want macros to be able to temporarily redefine : to do something
else without changing all the other logic around evaluation.

Basically, the choice was between having anything that needs to assign a
specific operation to : do a manual walk of the syntax tree according to its
own very specific rules OR to parse : as a call to _declare and thus having
everything that wants to assign a special meaning to : do it implicitly by
redefining _declare.

In the end I decided that parsing it as a call is more generic and more
flexible.