builtin/fun: define _define for functions
This moves towards a fix for a conceptual bug where we had the same macro
for definitions at the top-level (which creates globals) and definitions
inside functions (which should create locals).

We can split it up into two macros and define the one creating locals only
inside builtin_macro_fun.