elf: generate trampoline to entry point
We cannot set the entry point directly at the function given by the user
as the stack doesn't contain a return value for us to return to. So we
instead generate a new function with a call to the entry point given by
the user and finish off with a syscall to terminate the process.