function: always output to object
Instead of having the function itself keep track of output bytes,
relocations, and (assembly) comments (and likely things like debug
info in the future), we can reuse the object class which already
holds this info for code generated for the target.

One thing I don't particularly like about this is the double
indirection to generate code now, i.e. 'struct function' used to
directly embed a std::vector, but now it needs to dereference
'struct object' to emit a single byte. We'll work on that later.
(This may make this commit seem counterproductive, but it's more
helpful that we are now able to reuse the same infrastructure for
e.g. disassembly rather than having it only for host functions.)