compile: evaluate literals as values by default
This means you can directly do

    x := "foo";

instead of

    x := str "foo";

This is currently done by a direct use of the corresponding type constructors
rather than going through e.g. a named macro. We may want to do that in the
future to allow defaulting to constructing different types, but for now this
is good enough.