ast: get rid of std::shared_ptr
This is a pretty big rework of the AST data structure.

I essentially wanted to replace the user of std::shared_ptr with something
much less heavyweight and pointer-intensive, so what I've ended up with is
having a vector of AST nodes and then using indices to point to other nodes
in the tree.

The parsing speed is improved by ~80%. Text size is reduced by ~122 KiB.
Total heap allocs is reduced by ~76%, but total heap bytes allocated
increases (probably because std::vector reserves more than is actually
used).

There are two main dangers:

 - retaining a pointer to an AST node across std::vector modifications
   (e.g. pointer invalidation)

 - attempting to dereference AST node from one tree using the nodes of
   another tree