ast: factor out literal data
Instead of converting literal data to an internal representation at parse
time, we can do this during compilation instead.

This saves us time and space, just look at the valgrind before/after stats
for a --no-compile run:

     HEAP SUMMARY:
         in use at exit: 408 bytes in 6 blocks
    -  total heap usage: 12,328 allocs, 12,322 frees, 201,619,619 bytes allocated
    +  total heap usage: 68 allocs, 62 frees, 100,739,536 bytes allocated

Of course, we will have to make some of those allocations during
compilation instead (and possibly multiple times if evaluating the same
parse tree multiple times), but I think it's still a net benefit not to
have to reallocate them all the time when resizing the vector of AST
nodes.

There is about an 18% speedup for the same run.

Arguably there is also some benefit in potentially being able to catch
and report syntax errors slightly faster.

This also removes the dependency on std::string and libgmp for the parser
itself.

A small downside is that we need helpers to do the conversion instead of
just accessing the data directly.