builtin/elf: constructs segments in-memory before writing out to file
In order to more easily support more segments, we can calculate all
segment and symbol addresses ahead of time. It is not actually necessary
to construct the full segments in memory to do this, but it's convenient
as we can keep all the logic in one place.

Another reason to do this is that we used to just write out objects one
by one and that can potentially end up doing a bazillion small write()
calls. (This is surely something to revisit in the future, but for now
we don't have large enough programs to do useful benchmarking.)