object: get rid of std::map for comments
We used to store comments as a std::map<offset, std::vector<...>>, which
was gnarly to read but also probably way less efficient than it should be.

Let's just create a new type, 'comment', and use a plain vector. We can
walk it linearly as we disassemble the assembly code, so there's really
no reason to use std::map.