bytecode: new emitter + interpreter
Instead of JIT-compiling to x86-64 and executing the code directly for
eval(), we emit bytecode and interpret it. The benefit is that it will
allow us to cross-compile from any host (i.e. we are not relying on the
host being x86-64). Note that we are still JIT-compiling for functions
and macros (including those that end up in an executable), so this is
just a first step.