ast: put strings back in AST
There were two drawbacks in the old/current design:

1. Splitting the parsing of the string in two (parser + compiler) is bad
   when we can do it in one place.

2. If we parse a file once and compile it twice, we still need to parse
   the string twice.

So instead we keep an array of strings and the node just has an index into
that array (just like we do for child nodes).

We may want to do something similar for integer and floating-point
literals, but let's focus on more important things for now.